Reorganisation des fichiers, amélioration de la lisibilite
