--- src/Decomposition.h
@@ -1,25 +1,0 @@
-
-
-
-
-class Decomposition
-{
-
-
-
-
-    private :
-
-    FormulationPricer* p_formulation_pricer;
-    FormulationMaster* p_formulation_master;
-    ObjPricerUCP* p_obj_pricer_ucp;
-    VariableMaster* p_variable_master;
-
-
-
-
-}
-
-
-
-
