+++ src/UnitDecomposition/FormulationPricerUnitDecomposition.h
@@ -1,25 +1,36 @@
 /** 
  * @class FormulationPricerUnitDecomposition
- ** An object from this class has the variable and constraints defining a UCP problem.
- *  todo : improve the printing to get it with cout ?
 */
 
 #ifndef FormulationPricerUnitDecomposition_H
 #define FormulationPricerUnitDecomposition_H
 
+//** Includes
 
-/* standart includes */
+//* Standart
 #include <vector>
 #include <iostream>
 #include <string>
 #include <fstream>
+#include <sstream>
 
-/* SCIP includes */
+
+//* SCIP
+#include <scip/scipdefplugins.h>
 #include <scip/scip.h>
 
-/* user includes */
-#include "InstanceUCP.h"
-#include "../Decomposition/FormulationPricer.h"
+
+//* User
+
+// general
+#include "DataClasses/InstanceUCP.h"
+#include "DataClasses/ProductionPlan.h"
+
+// Decomposition
+#include "Decomposition/FormulationPricer.h"
+
+// Unit Decomposition
+#include "UnitDecomposition/FormulationPricerUnitDecomposition.h"
 
 
 /* namespace */
